Re-organising repo - starting with digital and SPI.
