+++ Cargo.toml
@@ -1,28 +1,6 @@
-[package]
-authors = [
-  "Jorge Aparicio <jorge@japaric.io>",
-  "Jonathan 'theJPster' Pallant <github@thejpster.org.uk>"
+[workspace]
+members = [
+	"embedded-hal",
+	"embedded-hal-digital",
+	"embedded-hal-spi",
 ]
-categories = ["asynchronous", "embedded", "hardware-support", "no-std"]
-description = " A Hardware Abstraction Layer (HAL) for embedded systems "
-documentation = "https://docs.rs/embedded-hal"
-edition = "2018"
-keywords = ["hal", "IO"]
-license = "MIT OR Apache-2.0"
-name = "embedded-hal"
-readme = "README.md"
-repository = "https://github.com/rust-embedded/embedded-hal"
-version = "0.2.3"
-
-[dependencies.nb]
-version = "0.1.1"
-
-[dev-dependencies]
-stm32f3 = { version = "0.8", features = ["stm32f303", "rt"] }
-futures = "0.1.17"
-
-[features]
-unproven = ["nb/unstable"]
-
-[package.metadata.docs.rs]
-features = ["unproven"]
