+++ embedded-hal-spi/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "embedded-hal-spi"
+version = "1.0.0-rc1"
+authors = ["Rust Embedded Working Group <hal@teams.rust-embedded.org>"]
+edition = "2018"
+categories = ["gpio", "embedded", "hardware-support", "no-std"]
+description = " A Serial Peripheral Interface (SPI) Hardware Abstraction Layer (HAL) for embedded systems "
+documentation = "https://docs.rs/embedded-hal-spi"
+keywords = ["hal", "IO", "gpio"]
+license = "MIT OR Apache-2.0"
+readme = "README.md"
+repository = "https://github.com/rust-embedded/embedded-hal"
+
+[dependencies]
+nb = "0.1"
